uniout now detects ipython and support ipython now
