+++ uniout.py
@@ -3,36 +3,18 @@
 
 __version__ = '0.2.2'
 
-import re
 import sys
+from _uniout import uniout
 
-# the helpers
 
-escape_x_re = re.compile(r'(?:\\x[0-9a-f]{2})+')
-escape_u_re = re.compile(r'(?:\\u[0-9a-f]{4}|\\U[0-9a-f]{8})+')
-encoding = sys.getfilesystemencoding()
+def runs_in_ipython():
 
-def dexuescape(s):
-    r'''decode the \x, \u and \U in a escaped string -> encoded string'''
-    s = escape_x_re.sub(lambda m: m.group().decode('string-escape'), s)
-    s = escape_u_re.sub(lambda m: m.group().decode('unicode-escape').encode(encoding), s)
+    import __builtin__
+    return '__IPYTHON__' in __builtin__.__dict__ and \
+           __builtin__.__dict__['__IPYTHON__']
 
-    # for Python < 2.7
-    if isinstance(s, unicode):
-        s = s.encode(encoding)
-
-    return s
-
-# make uniout
-uniout = lambda: 'middleware of stdout' # any instance
-
-# make uniout look like stdout
-for attrname in dir(sys.stdout):
-    if not attrname.startswith('__'):
-        setattr(uniout, attrname, getattr(sys.stdout, attrname))
-
-# modify the write method to de-escape
-uniout.write = lambda s: sys.__stdout__.write(dexuescape(s))
-
-# install the uniout
-sys.stdout = uniout
+if runs_in_ipython():
+    from IPython.utils import io
+    io.stdout = io.IOStream(uniout)
+else:
+    sys.stdout = uniout
